- cleaned up code a bit - added more detailed error handling - refactored some code into modules

+++ software/com-telemetry/src/test/test.c
@@ -42,7 +42,7 @@
         }
     }
 
-    void assertIntAreEqual(uint8_t expected, uint8_t actual, uint32_t lineNumber)
+    void assertIntAreEqual(uint32_t expected, uint32_t actual, uint32_t lineNumber)
     {
         if (expected != actual)
         {
